Added: migration data base with FlyAwayDB user profile page now users can change their email and password

+++ src/main/java/com/example/quepasa/domain/Message.java
@@ -1,14 +1,12 @@
 package com.example.quepasa.domain;
 
-import org.springframework.beans.factory.annotation.Autowired;
-
 import javax.persistence.*;
 
 @Entity
 public class Message {
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String text;
     private String tag;
@@ -49,11 +47,11 @@ public class Message {
         return text;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
